% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{edr_collections}
\alias{edr_collections}
\title{List collections offered by the service}
\usage{
edr_collections(client)
}
\arguments{
\item{client}{An \code{edr_client}.}
}
\value{
A tibble with one row per collection. Always includes \code{id},
\code{title}, \code{description}, \code{extent_bbox}, \code{crs}, \code{data_queries}, and
\code{links} columns.
}
\description{
List collections offered by the service
}
