% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr4r-package.R
\docType{package}
\name{edr4r-package}
\alias{edr4r}
\alias{edr4r-package}
\title{edr4r: Client for OGC API - Environmental Data Retrieval}
\description{
An R client for OGC API - EDR endpoints. Designed against the
Western Water Datahub (WWDH) pygeoapi deployment, but works with any
compliant EDR service. Build a client with \code{\link[=edr_client]{edr_client()}}, discover
resources with \code{\link[=edr_collections]{edr_collections()}}, and query data with
\code{\link[=edr_locations]{edr_locations()}}, \code{\link[=edr_area]{edr_area()}}, \code{\link[=edr_cube]{edr_cube()}}, \code{\link[=edr_position]{edr_position()}}, and
related verbs. Parse responses with \code{\link[=covjson_to_tibble]{covjson_to_tibble()}} and
\code{\link[=geojson_to_sf]{geojson_to_sf()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/internetofwater/WWDH}
  \item Report bugs at \url{https://github.com/internetofwater/WWDH/issues}
}

}
\author{
\strong{Maintainer}: Western Water Datahub \email{bwebb@lincolninst.edu}

}
\keyword{internal}
