% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{edr_collection}
\alias{edr_collection}
\title{Get a single collection's metadata}
\usage{
edr_collection(client, collection_id)
}
\arguments{
\item{client}{An \code{edr_client}.}

\item{collection_id}{Collection identifier, e.g. \code{"rise-edr"}.}
}
\value{
A list with the raw collection document.
}
\description{
Get a single collection's metadata
}
